+++ src/Assets.php
@@ -37,12 +37,21 @@ class Assets {
 	 * @param string|bool $locale   The locale to use for the KEC integration. Defaults to using the browser locale. Optional.
 	 */
 	public function __construct( $settings, $locale = false ) {
-		$this->assets_path = plugin_dir_url( __FILE__ ) . '../assets/';
+		$this->assets_path = self::get_assets_path();
 		$this->settings    = $settings;
 		$this->locale      = $locale;
 
 		add_action( 'init', array( $this, 'register_assets' ) );
 		add_action( 'wp_enqueue_scripts', array( $this, 'enqueue_assets' ), 15 );
+	}
+
+	/**
+	 * Get the assets folder path.
+	 *
+	 * @return string
+	 */
+	public static function get_assets_path() {
+		return plugin_dir_url( __DIR__ ) . 'assets/';
 	}
 
 	/**
@@ -84,6 +93,12 @@ class Assets {
 		$is_product_page = is_product();
 		$product         = $is_product_page ? wc_get_product( get_the_ID() ) : null;
 
+		$client_id = $this->settings->get_credentials_secret();
+
+		if ( empty( $client_id ) ) { // Skip if we don't have a client ID.
+			return;
+		}
+
 		$params = array(
 			'ajax'            => array(
 				'get_payload'   => array(
@@ -107,7 +122,7 @@ class Assets {
 				'id'   => $product->get_id(),
 				'type' => $product->get_type(),
 			) : null,
-			'client_id'       => $this->settings->get_credentials_secret(),
+			'client_id'       => $client_id,
 			'theme'           => $this->settings->get_theme(),
 			'shape'           => $this->settings->get_shape(),
 			'locale'          => $this->locale,