+++ src/Settings.php
@@ -73,8 +73,9 @@ class Settings {
 		return array(
 			'kec_enabled'            => $settings['kec_enabled'] ?? 'no',
 			'kec_credentials_secret' => $settings['kec_credentials_secret'] ?? '',
-			'kec_theme'              => $settings['kec_theme'] ?? 'default',
+			'kec_theme'              => $settings['kec_theme'] ?? 'dark',
 			'kec_shape'              => $settings['kec_shape'] ?? 'default',
+			'kec_placement'          => $settings['kec_placement'] ?? 'both',
 		);
 	}
 
@@ -93,6 +94,10 @@ class Settings {
 	 * @return string
 	 */
 	public function get_credentials_secret() {
+		if ( function_exists( 'kp_get_client_id_by_currency' ) ) {
+			return kp_get_client_id_by_currency();
+		}
+
 		return $this->options['kec_credentials_secret'] ?? '';
 	}
 
@@ -102,7 +107,7 @@ class Settings {
 	 * @return string
 	 */
 	public function get_theme() {
-		return $this->options['kec_theme'] ?? 'default';
+		return $this->options['kec_theme'] ?? 'dark';
 	}
 
 	/**
@@ -115,58 +120,109 @@ class Settings {
 	}
 
 	/**
+	 * Get the placements for the Klarna Express Checkout.
+	 *
+	 * @return string
+	 */
+	public function get_placements() {
+		return $this->options['kec_placement'] ?? 'both';
+	}
+
+	/**
 	 * Get the setting fields.
 	 *
 	 * @return array
 	 */
 	public function get_setting_fields() {
-		$portal_live_link = '<a href="https://portal.klarna.com/" target="_blank">' . __( 'here for production', 'klarna-express-checkout' ) . '</a>';
-		$portal_test_link = '<a href="https://portal.playground.klarna.com/" target="_blank">' . __( 'here for playground', 'klarna-express-checkout' ) . '</a>';
-		// translators: %1$s is the link to the Klarna Merchant Portal for production, %2$s is the link to the Klarna Merchant Portal for playground.
-		$credentials_secret_desc = sprintf( __( 'Enter your Klarna Client Identifier. This can be found in the Klarna Merchant Portal %1$s and %2$s.', 'klarna-express-checkout' ), $portal_live_link, $portal_test_link );
-
 		return array(
-			'kec_settings'           => array(
-				'title' => __( 'Klarna Express Checkout', 'klarna-express-checkout' ),
-				'type'  => 'title',
-				'desc'  => __( 'Klarna Express Checkout is a fast and easy way for customers to pay with Klarna.', 'klarna-express-checkout' ),
+			'kec_settings'  => array(
+				'id'          => 'kec_settings',
+				'title'       => 'Express Checkout',
+				'description' => __( 'Offer a faster check-out process that will lower the threshold for shoppers to complete a purchase.', 'klarna-express-checkout' ),
+				'links'       => array(
+					array(
+						'url'   => 'https://docs.klarna.com/express-checkout/',
+						'title' => __( 'Documentation', 'klarna-express-checkout' ),
+					),
+				),
+				'type'        => 'kp_section_start',
 			),
-			'kec_enabled'            => array(
+			'kec_enabled'   => array(
 				'title'   => __( 'Enable/Disable', 'klarna-express-checkout' ),
 				'type'    => 'checkbox',
 				'label'   => __( 'Enable Klarna Express Checkout', 'klarna-express-checkout' ),
 				'default' => 'no',
 			),
-			'kec_credentials_secret' => array(
-				'title'       => __( 'Klarna Client Identifier', 'klarna-express-checkout' ),
-				'type'        => 'text',
-				'description' => $credentials_secret_desc,
-				'desc_tip'    => false,
+			'kec_info'      => array(
+				'type'        => 'kp_text_info',
+				'title'       => __( 'Placements & button style', 'klarna-express-checkout' ),
+				'description' => __( 'Tailor the express checkout button to fit your brand by adjusting the button theme, shape and selecting placements.', 'klarna-express-checkout' ),
 			),
-			'kec_theme'              => array(
+			'kec_theme'     => array(
 				'title'       => __( 'Theme', 'klarna-express-checkout' ),
 				'type'        => 'select',
 				'description' => __( 'Select the theme for the Klarna Express Checkout.', 'klarna-express-checkout' ),
-				'desc_tip'    => false,
+				'desc_tip'    => true,
 				'options'     => array(
-					'default' => __( 'Default', 'klarna-express-checkout' ),
-					'dark'    => __( 'Dark', 'klarna-express-checkout' ),
-					'light'   => __( 'Light', 'klarna-express-checkout' ),
+					'dark'     => __( 'Dark', 'klarna-express-checkout' ),
+					'light'    => __( 'Light', 'klarna-express-checkout' ),
+					'outlined' => __( 'Outlined', 'klarna-express-checkout' ),
 				),
-				'default'     => 'default',
+				'default'     => 'dark',
 			),
-			'kec_shape'              => array(
+			'kec_shape'     => array(
 				'title'       => __( 'Shape', 'klarna-express-checkout' ),
 				'type'        => 'select',
 				'description' => __( 'Select the shape for the Klarna Express Checkout.', 'klarna-express-checkout' ),
-				'desc_tip'    => false,
+				'desc_tip'    => true,
 				'options'     => array(
-					'default' => __( 'Default', 'klarna-express-checkout' ),
+					'default' => __( 'Rounded', 'klarna-express-checkout' ),
 					'rect'    => __( 'Rectangular', 'klarna-express-checkout' ),
 					'pill'    => __( 'Pill', 'klarna-express-checkout' ),
 				),
 				'default'     => 'default',
 			),
+			'kec_placement' => array(
+				'title'   => __( 'Placements', 'klarna-express-checkout' ),
+				'type'    => 'select',
+				'default' => 'both',
+				'options' => array(
+					'both'    => __( 'All (recommended)', 'klarna-express-checkout' ),
+					'product' => __( 'Product pages', 'klarna-express-checkout' ),
+					'cart'    => __( 'Cart page', 'klarna-express-checkout' ),
+				),
+			),
+			'kec_end'       => array(
+				'type'     => 'kp_section_end',
+				'previews' => array(
+					array(
+						'title' => __( 'Preview', 'klarna-express-checkout' ),
+						'image' => $this->get_preview_img_url(),
+					),
+				),
+			),
 		);
+	}
+
+	/**
+	 * Get the preview image url.
+	 *
+	 * @return string
+	 */
+	public function get_preview_img_url() {
+		$shape = $this->get_shape();
+		$theme = $this->get_theme();
+
+		if ( '' === $shape ) {
+			$shape = 'default';
+		}
+
+		if ( '' === $theme || 'default' === $theme ) {
+			$theme = 'dark';
+		}
+
+		$preview_img = Assets::get_assets_path() . 'img/preview-' . $shape . '-' . $theme . '.png';
+
+		return $preview_img;
 	}
 }