+++ tests/unit-tests/SettingsTest.php
@@ -16,8 +16,9 @@ class SettingsTest extends TestCase {
 				array(
 					'kec_enabled'            => 'yes',
 					'kec_credentials_secret' => 'test_credentials_secret',
-					'kec_theme'              => 'default',
+					'kec_theme'              => 'dark',
 					'kec_shape'              => 'default',
+					'kec_placement'          => 'both',
 				)
 			);
 		$this->settings = new Settings( 'test_key' );
@@ -37,28 +38,32 @@ class SettingsTest extends TestCase {
 	}
 
 	public function testGetTheme() {
-		$this->assertEquals( 'default', $this->settings->get_theme() );
+		$this->assertEquals( 'dark', $this->settings->get_theme() );
 	}
 
 	public function testGetShape() {
 		$this->assertEquals( 'default', $this->settings->get_shape() );
 	}
 
+	public function testGetPlacement() {
+		$this->assertEquals( 'both', $this->settings->get_placements() );
+	}
+
 	public function testGetSettings() {
 		$result = $this->settings->get_settings();
 
-		$this->assertArrayHasKey( 'kec_credentials_secret', $result );
 		$this->assertArrayHasKey( 'kec_theme', $result );
 		$this->assertArrayHasKey( 'kec_shape', $result );
+		$this->assertArrayHasKey( 'kec_placement', $result );
 	}
 
 	public function testGetSettingFields() {
 		$result = $this->settings->get_setting_fields();
 
 		$this->assertArrayHasKey( 'kec_settings', $result );
-		$this->assertArrayHasKey( 'kec_credentials_secret', $result );
 		$this->assertArrayHasKey( 'kec_theme', $result );
 		$this->assertArrayHasKey( 'kec_shape', $result );
+		$this->assertArrayHasKey( 'kec_placement', $result );
 	}
 
 	public function testAddSettings() {
@@ -70,7 +75,7 @@ class SettingsTest extends TestCase {
 
 		$this->assertArrayHasKey( 'test_setting', $newSettings );
 		$this->assertArrayHasKey( 'kec_settings', $newSettings );
-		$this->assertArrayHasKey( 'kec_credentials_secret', $newSettings );
 		$this->assertArrayHasKey( 'kec_theme', $newSettings );
+		$this->assertArrayHasKey( 'kec_placement', $newSettings );
 	}
 }