+++ changelog.md
@@ -7,12 +7,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
 ------------------
+
+## [1.4.0] - 2024-09-11
+### Added
+
+* Preview images for each of the button theme and shape combinations.
+* A new theme for the outlined version of the KEC button.
+
+### Changed
+
+* Changed the settings to match the new layout and structure for the Klarna Payments plugins settings pages. Changed the WooCommerce settings API section start and end to be of the types `kp_section_start` and `kp_section_end` respectively.
+
+### Removed
+
+* Removed the setting for the Client ID from the package, and instead uses the new `kp_get_client_id` function from Klarna Payments to get the client id based on the customers country and availability.
+* Removed the default theme option, since this was just a different name for the dark theme.
+
+### Fixed
+
+* Fixed an issue with showing the Klarna Express Checkout button on the product page when the product is out of stock.
+
 ## [1.3.1] - 2024-04-22
 ### Fix
 
-* Fixed an issue where a new KEC button would be created on a variation product every time the customer selected a variation option. 
+* Fixed an issue where a new KEC button would be created on a variation product every time the customer selected a variation option.
 * Moved from client_key to client_id as the former has now been deprecated by Klarna.
 
 ## [1.3.0] - 2024-01-31